cleaning up um + db avatar controllers + will add get avatar file too
