+++ server/controllers/dev/view.controller.ts
@@ -9,3 +9,13 @@ export const viewDB = async (request: FastifyRequest, reply: FastifyReply): Prom
 	}
 	return await res.json();
 };
+
+export const viewID = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
+	const { id } = request.query as { id: string };
+
+	const res = await fetch(`${DATABASE_URL}/viewID/${id}`);
+	if (!res.ok) {
+	throw new Error(`Failed to fetch user with id ${id}: ${res.statusText}`);
+	}
+	return await res.json();
+};
