+++ server/controllers/user/avatar.controller.ts
@@ -1,6 +1,4 @@
 import { FastifyRequest, FastifyReply } from 'fastify';
-
-// for editing/uploading avatar
 import FormData from 'form-data';
 import fetch from 'node-fetch';
 const DATABASE_URL = 'http://database_container:3000';
