+++ server/controllers/user/edit.controller.ts
@@ -67,3 +67,25 @@ export const editEmail = async (request: FastifyRequest, reply: FastifyReply): P
 		return reply.code(err.code).send({ error: err.message });
 	}
 };
+
+export const deleteUser = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
+	try {
+		const { username } = request.params as { username: string };
+
+		const res = await fetch(`${DATABASE_URL}/delete/${username}`, {
+			method: 'DELETE',
+			headers: { 'Content-Type': 'application/json' },
+			body: JSON.stringify({ username }),
+		});
+		if (!res.ok) {
+			const responseBody = await res.json() as { error: string };
+			throw { code: res.status, message: responseBody.error };
+		}
+		return ({ code: res.status });
+	} catch (error) {
+		console.error(error);
+		const err = error as { code: number, message: string };
+		return reply.code(err.code).send({ error: err.message });
+	}
+};
+
