--- server/controllers/user/logout.controller.ts
@@ -1,20 +1,0 @@
-import { FastifyRequest, FastifyReply } from 'fastify';
-
-const DATABASE_URL = 'http://database_container:3000';
-
-export const logout = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
-	try {
-		const { username } = request.params as { username: string };
-		
-		const res = await fetch(`${DATABASE_URL}/logout/${username}`);
-		if (!res.ok) {
-			const responseBody = await res.json() as { error: string };
-			throw { code: res.status, message: responseBody.error };
-		}
-		return ({ code: res.status });
-	} catch (error) {
-		console.error(error);
-		const err = error as { code: number, message: string };
-		return reply.code(err.code).send({ error: err.message });
-	}
-};
