+++ server/controllers/user/register.controller.ts
@@ -22,3 +22,41 @@ export const register = async (request: FastifyRequest, reply: FastifyReply): Pr
 		return reply.code(err.code).send({ error: err.message });
 	}
 };
+
+export const login = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
+	try {
+		const { username, password } = request.body as { username: string, password: string };
+		
+		const res = await fetch(`${DATABASE_URL}/login`, {
+			method: 'POST',
+			headers: { 'Content-Type': 'application/json' },
+			body: JSON.stringify({ username, password }),
+		});
+		if (!res.ok) {
+			const responseBody = await res.json() as { error: string };
+			throw { code: res.status, message: responseBody.error };
+		}
+		return ({ code: res.status });
+	} catch (error) {
+		console.error(error);
+		const err = error as { code: number, message: string };
+		return reply.code(err.code).send({ error: err.message });
+	}
+};
+
+export const logout = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
+	try {
+		const { username } = request.params as { username: string };
+		
+		const res = await fetch(`${DATABASE_URL}/logout/${username}`);
+		if (!res.ok) {
+			const responseBody = await res.json() as { error: string };
+			throw { code: res.status, message: responseBody.error };
+		}
+		return ({ code: res.status });
+	} catch (error) {
+		console.error(error);
+		const err = error as { code: number, message: string };
+		return reply.code(err.code).send({ error: err.message });
+	}
+};
